requirements.txt
PySide6
openpyxl
Pillow
PyMuPDF
reportlab
scikit-image
numpy
pyinstaller
